# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Task Resource OpenApi","contact":{"name":"\u5c0f\u73a9\u610f\u513f","url":"http:\/\/140.143.186.207:8081\/shopping\/public\/api\/documentation","email":"support@todo.test"},"version":"3.0"},"servers":[{"url":"http:\/\/140.143.186.207:8081\/api\/"}],"paths":{"\/admin":{"get":{"tags":["\u540e\u53f0\u7528\u6237\u4fe1\u606f\u6a21\u5757"],"summary":"\u83b7\u53d6\u767b\u5f55\u7528\u6237\u7684\u4fe1\u606f","description":"\u83b7\u53d6\u767b\u5f55\u7528\u6237\u7684\u4fe1\u606f","operationId":"getAdminSession","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"put":{"tags":["\u540e\u53f0\u7528\u6237\u4fe1\u606f\u6a21\u5757"],"summary":"\u4fee\u6539\u7528\u6237\u4fe1\u606f","description":"create new task","operationId":"updateAdmin","parameters":[{"name":"name","in":"query","description":"\u7528\u6237\u540d","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","description":"email","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/user":{"get":{"tags":["\u524d\u53f0\u7528\u6237\u4fe1\u606f\u6a21\u5757"],"summary":"\u83b7\u53d6\u767b\u5f55\u7528\u6237\u7684\u4fe1\u606f","description":"\u83b7\u53d6\u767b\u5f55\u7528\u6237\u7684\u4fe1\u606f","operationId":"getUsersSession","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"put":{"tags":["\u524d\u53f0\u7528\u6237\u4fe1\u606f\u6a21\u5757"],"summary":"\u4fee\u6539\u7528\u6237\u4fe1\u606f","description":"create new task","operationId":"updateUser","parameters":[{"name":"name","in":"query","description":"\u7528\u6237\u540d","required":false,"schema":{"type":"string"}},{"name":"tel","in":"query","description":"\u7535\u8bdd\u53f7\u7801","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/manage\/login":{"get":{"tags":["\u540e\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u767b\u5f55","description":"\u7528\u6237\u767b\u5f55","operationId":"login","parameters":[{"name":"name","in":"query","description":"\u7528\u6237\u540d","required":false,"schema":{"type":"String"}},{"name":"password","in":"query","description":"\u5bc6\u7801","required":false,"schema":{"type":"String"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/manage\/loginOut":{"get":{"tags":["\u540e\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u6ce8\u9500","description":"\u7528\u6237\u6ce8\u9500","operationId":"loginOut","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/manage\/register":{"post":{"tags":["\u540e\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u6ce8\u518c","description":"create new task","operationId":"insert","parameters":[{"name":"name","in":"query","description":"\u540d\u5b57","required":false,"schema":{"type":"String"}},{"name":"password","in":"query","description":"\u5bc6\u7801","required":false,"schema":{"type":"String"}},{"name":"email","in":"query","description":"\u90ae\u4ef6","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/manage\/product":{"get":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u83b7\u53d6\u5168\u90e8\u5546\u54c1\u4fe1\u606f","description":"\u5546\u54c1\u7ba1\u7406\u6a21\u5757","operationId":"index","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"put":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u4fee\u6539\u5546\u54c1\u4fe1\u606f","description":"create new task","operationId":"update","parameters":[{"name":"id","in":"query","description":"idid","required":false,"schema":{"type":"int"}},{"name":"name","in":"query","description":"\u5546\u54c1\u540d\u79f0","required":false,"schema":{"type":"string"}},{"name":"price","in":"query","description":"\u4ef7\u683c","required":false,"schema":{"type":"int"}},{"name":"number","in":"query","description":"\u6570\u91cf\u5e93\u5b58","required":false,"schema":{"type":"int"}},{"name":"describe","in":"query","description":"\u5546\u54c1\u63cf\u8ff0","required":false,"schema":{"type":"String"}},{"name":"typeId","in":"query","description":"\u5546\u54c1\u5206\u7c7b,\u4e3anull\u65f6\u4e0d\u4fee\u6539typeId","required":false,"schema":{"type":"int"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]},"post":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u6dfb\u52a0\u5546\u54c1\u4fe1\u606f","description":"create new task","operationId":"insert","parameters":[{"name":"name","in":"query","description":"\u5546\u54c1\u540d\u79f0","required":false,"schema":{"type":"string"}},{"name":"price","in":"query","description":"\u4ef7\u683c","required":false,"schema":{"type":"int"}},{"name":"number","in":"query","description":"\u5e93\u5b58\u6570\u91cf,\u9ed8\u8ba4100\u4ef6","required":false,"schema":{"type":"int"}},{"name":"describe","in":"query","description":"\u5546\u54c1\u63cf\u8ff0","required":false,"schema":{"type":"String"}},{"name":"typeId","in":"query","description":"\u5546\u54c1\u5206\u7c7b","required":false,"schema":{"type":"int"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/manage\/product\/getByTypeId\/{typeId}":{"get":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9a\u5546\u54c1\u5206\u7c7b\u7684\u5546\u54c1","description":"","operationId":"getByTypeId","parameters":[{"name":"typeId","in":"path","description":"typeId","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/manage\/product\/{id}":{"get":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9aid\u7684\u5546\u54c1","description":"\u7528\u6237\u7ba1\u7406\u6a21\u5757","operationId":"getById","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"delete":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u5220\u9664\u6307\u5b9aid\u7684\u5546\u54c1","description":"\u5546\u54c1\u7ba1\u7406\u6a21\u5757\u7ba1\u7406\u6a21\u5757","operationId":"delete","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/manage\/product\/upload\/{id}":{"post":{"tags":["\u540e\u53f0\u5546\u54c1\u6a21\u5757"],"summary":"\u4e0a\u4f20\u5546\u54c1\u56fe\u7247","description":"\u5546\u54c1\u7ba1\u7406\u6a21\u5757","operationId":"upload","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}},{"name":"imgUrl","in":"path","description":"imgUrl","required":false,"schema":{"type":"file"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/manage\/type":{"get":{"tags":["\u540e\u53f0\u5546\u54c1\u5206\u7c7b\u6a21\u5757"],"summary":"\u83b7\u53d6\u5168\u90e8\u5546\u54c1\u5206\u7c7b\u4fe1\u606f","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"getAll","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"put":{"tags":["\u540e\u53f0\u5546\u54c1\u5206\u7c7b\u6a21\u5757"],"summary":"\u4fee\u6539\u5546\u54c1\u5206\u7c7b","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"update","parameters":[{"name":"id","in":"query","description":"idid","required":false,"schema":{"type":"int"}},{"name":"name","in":"query","description":"\u5546\u54c1\u5206\u7c7b\u540d\u79f0","required":false,"schema":{"type":"string"}},{"name":"describe","in":"query","description":"\u5546\u54c1\u5206\u7c7b\u63cf\u8ff0","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]},"post":{"tags":["\u540e\u53f0\u5546\u54c1\u5206\u7c7b\u6a21\u5757"],"summary":"\u6dfb\u52a0\u5546\u54c1\u5206\u7c7b","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"post","parameters":[{"name":"name","in":"query","description":"\u5546\u54c1\u5206\u7c7b\u540d\u79f0","required":false,"schema":{"type":"string"}},{"name":"describe","in":"query","description":"\u5546\u54c1\u5206\u7c7b\u63cf\u8ff0","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/manage\/type\/{id}":{"get":{"tags":["\u540e\u53f0\u5546\u54c1\u5206\u7c7b\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9aid\u7684\u5546\u54c1\u5206\u7c7b","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"get","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"delete":{"tags":["\u540e\u53f0\u5546\u54c1\u5206\u7c7b\u6a21\u5757"],"summary":"\u5220\u9664\u6307\u5b9aid\u7684\u5206\u7c7b","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"delete","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/address\/getByUserId":{"get":{"tags":["\u524d\u53f0\u6536\u8d27\u5730\u5740\u6a21\u5757"],"summary":"\u83b7\u53d6\u767b\u5f55\u8d26\u6237\u7684\u5730\u5740\u4fe1\u606f","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"getByUserId","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/address\/{id}":{"get":{"tags":["\u524d\u53f0\u6536\u8d27\u5730\u5740\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9aid\u7684\u5730\u5740\u4fe1\u606f","description":"\u6536\u8d27\u5730\u5740\u6a21\u5757","operationId":"getById","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"delete":{"tags":["\u524d\u53f0\u6536\u8d27\u5730\u5740\u6a21\u5757"],"summary":"\u5220\u9664\u6307\u5b9aid\u7684\u5730\u5740\u4fe1\u606f","description":"\u6536\u8d27\u5730\u5740\u6a21\u5757","operationId":"delete","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/address":{"put":{"tags":["\u524d\u53f0\u6536\u8d27\u5730\u5740\u6a21\u5757"],"summary":"\u4fee\u6539\u6307\u5b9aid\u7684\u5730\u5740","description":"\u6536\u8d27\u5730\u5740\u6a21\u5757","operationId":"update","parameters":[{"name":"id","in":"query","description":"IDID","required":false,"schema":{"type":"integer"}},{"name":"name","in":"query","description":"\u6536\u4ef6\u4eba\u59d3\u540d","required":false,"schema":{"type":"string"}},{"name":"tel","in":"query","description":"\u7535\u8bdd\u53f7\u7801","required":false,"schema":{"type":"string"}},{"name":"address","in":"query","description":"\u5730\u5740","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"post":{"tags":["\u524d\u53f0\u6536\u8d27\u5730\u5740\u6a21\u5757"],"summary":"\u6dfb\u52a0\u5730\u5740","description":"\u6536\u8d27\u5730\u5740\u6a21\u5757","operationId":"insert","parameters":[{"name":"name","in":"query","description":"\u6536\u4ef6\u4eba\u59d3\u540d","required":false,"schema":{"type":"string"}},{"name":"tel","in":"query","description":"\u7535\u8bdd\u53f7\u7801","required":false,"schema":{"type":"string"}},{"name":"address","in":"query","description":"\u5730\u5740","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/order":{"post":{"tags":["\u524d\u53f0\u8ba2\u5355\u6a21\u5757"],"summary":"\u65b0\u589e\u8ba2\u5355","description":"create new task","operationId":"insert","parameters":[{"name":"productIds","in":"query","description":"productIds","required":false,"schema":{"type":"String"}},{"name":"addressId","in":"query","description":"addressId","required":false,"schema":{"type":"String"}},{"name":"numbers","in":"query","description":"numbers","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/shopping\/order\/pay":{"put":{"tags":["\u524d\u53f0\u8ba2\u5355\u6a21\u5757"],"summary":"\u4e0b\u5355","description":"create new task","operationId":"update","parameters":[{"name":"id","in":"query","description":"\u8ba2\u5355id","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/shopping\/shoppingCart":{"get":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u83b7\u53d6\u5168\u90e8\u8be5\u7528\u6237\u5168\u90e8\u8d2d\u7269\u8f66\u4fe1\u606f","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"getAll","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"post":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u6dfb\u52a0\u5546\u54c1\u5230\u8d2d\u7269\u8f66","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"insert","parameters":[{"name":"productId","in":"query","description":"\u5546\u54c1id","required":false,"schema":{"type":"int"}},{"name":"number","in":"query","description":"\u52a0\u5165\u8d2d\u7269\u8f66\u7684\u5546\u54c1\u6570\u91cf,\u9ed8\u8ba4\u4e3a1","required":false,"schema":{"type":"int"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/shopping\/shoppingCart\/getById\/{id}":{"get":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9aid\u7684\u8d2d\u7269\u8f66\u4fe1\u606f","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"get","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/shoppingCart\/{id}":{"delete":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u5220\u9664\u6307\u5b9aid\u7684\u8d2d\u7269\u8f66\u5546\u54c1","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"delete","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/shoppingCart\/batchDelete\/{ids}":{"delete":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u6279\u91cf\u5220\u9664\u6307\u5b9aid\u7684\u8d2d\u7269\u8f66\u5546\u54c1","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"batchDelete","parameters":[{"name":"ids","in":"path","description":"1,2,3","required":false,"schema":{"type":"String"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/shoppingCart\/batchInsert":{"post":{"tags":["\u524d\u53f0\u8d2d\u7269\u8f66\u6a21\u5757"],"summary":"\u6279\u91cf\u6dfb\u52a0\u5546\u54c1\u5230\u8d2d\u7269\u8f66","description":"\u8d2d\u7269\u8f66\u6a21\u5757","operationId":"batchInsert","parameters":[{"name":"productIds","in":"query","description":"1,2,3","required":false,"schema":{"type":"String"}},{"name":"numbers","in":"query","description":"1,2,3","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/shopping":{"get":{"tags":["\u524d\u53f0\u5546\u54c1\u5c55\u793a\u6a21\u5757"],"summary":"\u83b7\u53d6\u5168\u90e8\u5546\u54c1\u4fe1\u606f","description":"\u5546\u54c1\u7ba1\u7406\u6a21\u5757","operationId":"index","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/type":{"get":{"tags":["\u524d\u53f0\u5546\u54c1\u5c55\u793a\u6a21\u5757"],"summary":"\u83b7\u53d6\u5168\u90e8\u5546\u54c1\u5206\u7c7b\u4fe1\u606f","description":"\u5546\u54c1\u5206\u7c7b\u6a21\u5757","operationId":"getTypeAll","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/getByTypeId\/{typeId}":{"get":{"tags":["\u524d\u53f0\u5546\u54c1\u5c55\u793a\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9a\u5546\u54c1\u5206\u7c7b\u7684\u5546\u54c1","description":"","operationId":"getByTypeId","parameters":[{"name":"typeId","in":"path","description":"typeId","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/{id}":{"get":{"tags":["\u524d\u53f0\u5546\u54c1\u5c55\u793a\u6a21\u5757"],"summary":"\u83b7\u53d6\u6307\u5b9aid\u7684\u5546\u54c1","description":"\u7528\u6237\u7ba1\u7406\u6a21\u5757","operationId":"getById","parameters":[{"name":"id","in":"path","description":"IDID","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/login":{"get":{"tags":["\u524d\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u767b\u5f55","description":"\u94b1\u5305\u7ba1\u7406\u6a21\u5757","operationId":"login","parameters":[{"name":"name","in":"query","description":"\u7528\u6237\u540d","required":false,"schema":{"type":"String"}},{"name":"password","in":"query","description":"\u5bc6\u7801","required":false,"schema":{"type":"String"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/loginOut":{"get":{"tags":["\u524d\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u6ce8\u9500","description":"\u7528\u6237\u6ce8\u9500","operationId":"loginOut","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/register":{"post":{"tags":["\u524d\u53f0\u7528\u6237\u767b\u5f55\u6ce8\u518c\u6a21\u5757"],"summary":"\u7528\u6237\u6ce8\u518c","description":"create new task","operationId":"insert","parameters":[{"name":"name","in":"query","description":"\u540d\u5b57","required":false,"schema":{"type":"String"}},{"name":"password","in":"query","description":"\u5bc6\u7801","required":false,"schema":{"type":"String"}},{"name":"tel","in":"query","description":"\u7535\u8bdd\u53f7\u7801","required":false,"schema":{"type":"String"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}},"\/shopping\/wallet":{"get":{"tags":["\u524d\u53f0\u94b1\u5305\u6a21\u5757"],"summary":"\u83b7\u53d6\u767b\u5f55\u7528\u6237\u7684\u94b1\u5305\u4fe1\u606f","description":"\u94b1\u5305\u7ba1\u7406\u6a21\u5757","operationId":"getWalletByUserId","responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]}},"\/shopping\/wallet\/addCredit":{"put":{"tags":["\u524d\u53f0\u94b1\u5305\u6a21\u5757"],"summary":"\u5145\u503c\u94b1\u5305","description":"create new task","operationId":"insert","parameters":[{"name":"money","in":"query","description":"\u91d1\u989d","required":false,"schema":{"type":"decimal"}}],"requestBody":{"request":"is_completed","description":"If the task is completed or not","required":true},"responses":{"201":{"description":"The task item created","content":{"application\/json":{"schema":{}}}}},"security":[{"passport":[]}]}}},"components":{"securitySchemes":{"passport":{"type":"oauth2","description":"Use a global client_id \/ client_secret and your email \/ password combo to obtain a token","in":"header","flows":{"password":{"authorizationUrl":"\/oauth\/authorize","tokenUrl":"\/oauth\/token","refreshUrl":"\/oauth\/token\/refresh","scopes":[]}},"scheme":"http"}}},"securityDefinitions":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"password":{"authorizationUrl":"http:\/\/localhost\/oauth\/authorize","tokenUrl":"http:\/\/localhost\/oauth\/token","refreshUrl":"http:\/\/localhost\/token\/refresh","scopes":[]}}}}}